// karabiner/types.ts
export interface KarabinerRules {
	description?: string;
	manipulators?: Manipulator[];
}

export interface Manipulator {
	description?: string;
	type: "basic";
	from: From;
	to?: To[];
	to_after_key_up?: To[];
	to_if_alone?: To[];
	parameters?: Parameters;
	conditions?: Conditions[];
}

export interface Parameters {
	"basic.simultaneous_threshold_milliseconds"?: number;
}

type Conditions =
	| FrontMostApplicationCondition
	| DeviceCondition
	| KeybaordTypeCondition
	| InputSourceCondition
	| VaribaleCondition
	| EventChangedCondition;
// ...
export interface From {
	key_code?: KeyCode;
	simultaneous?: SimultaneousFrom[];
	simultaneous_options?: SimultaneousOptions;
	modifiers?: Modifiers;
	pointing_button?: string;
}

export interface Modifiers {
	optional?: string[];
	mandatory?: string[];
}

export interface To {
	key_code?: KeyCode;
	modifiers?: string[];
	shell_command?: string;
	set_variable?: {
		name: string;
		value: boolean | number | string;
	};
	mouse_key?: MouseKey;
	pointing_button?: string;
	repeat?: boolean;
}